[knf/validators/system] Fix compatibility with osx
