+++ knf/validators/system/validators_test.go
@@ -91,26 +91,6 @@ func (s *ValidatorSuite) TestGroupValidator(c *C) {
 	c.Assert(errs, HasLen, 1)
 }
 
-func (s *ValidatorSuite) TestInterfaceValidator(c *C) {
-	configFile := createConfig(c, _CONFIG_DATA)
-
-	err := knf.Global(configFile)
-	c.Assert(err, IsNil)
-
-	errs := knf.Validate([]*knf.Validator{
-		{"interface:test0", Interface, nil},
-		{"interface:test1", Interface, nil},
-	})
-
-	c.Assert(errs, HasLen, 0)
-
-	errs = knf.Validate([]*knf.Validator{
-		{"interface:test2", Interface, nil},
-	})
-
-	c.Assert(errs, HasLen, 1)
-}
-
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func createConfig(c *C, data string) string {
